Updates for primo formats
Primo formats appears to have two new formats.  These break with existing
format strings in that (a) the facet versions appear to be singular, and
(b) there is no word separator.

These format strings are converted for human consumption via a mapping
maintained within the field configration.

Additionally, the Reference Entry format does not have an icon associated
with it.  The mapping of format, to icons is maintained in the
IconMetadataComponent class.

Previously used, and commented out field definitions have also been removed
from the fields configuration file.