LIBSEARCH-976
Implement a new availability filter for in Spectrum.
The solr field has 5 values in it, only 3 of which are to be
presented to the end-user.  However, they are also not presented
verbatim.

The mapping is:
physcial => Physical
hathi_trust_full_text_or_electronic_holding => Available online
hathi_trust_or_electronic_holding => Available online
hathi_trust => HathiTrust
hathi_trust_full_text => HathiTrust

Which 3 are presented depends on the value of the request's
search_only? method.
The hathi_trust_full_text variants are used when search_only? is true.
The other two are used when search_only? is false.

The same logic must be applied in reverse when generating a query for
solr.