Adding prometheus metrics
This may not be the final configuration.

In particular, I don't fully understand why Prometheus::Client.config.data_store
has to be set in both config.ru and config/puma.rb.  Some metrics
appear to be absent if I don't. That said.  metrics are being served in my
docker-compose environment, so I think this is a good starting point.