Adding a redis-based request throttle.
In case we need it in the future, I'm writing out an example of
a throttle using redis.  Not using redis would mean each worker
has its own counter, and we have enough workers for that to be
a problem.

Environment variables introduced are:
RACK_ATTACK_REDIS_URL
RACK_ATTACK_THROTTLE_LIMIT
RACK_ATTACK_THROTLE_PERIOD