More http_server_request_duration_seconds buckets
Add more buckets to http_server_request_duration_seconds.
10 seconds is unfortunately not long enough.

The weirdness with the Prometheus::Middleware::Collector
initializing the registry is avoided by creating a
subclass that assumes a the registry is already
populated, and fetches the metrics object from it.

The exceptions were exporting malformed metrics.  Rather
than redefining the #trace method, the RegisteredCollector
defaults to an object that shares enough of the same interface
 as a metric to fool #trace and #record.  Because the FakeMetric
is not part of the registry, the exporter does not attempt to
export them.