Refactor the page/page_size computation
Primo and Summon use a page number and page size for pagination.
Spectrum and Solr use an offset and count.
Thus we need to compute the page number and page size for the search.
Spectrum supports arbitrary values for page number and page size, thus
in order to fetch the results in a single request we fetch a range of
records and return a slice of the results.
For example, start=45 and count=10 could be met with page_size=15,
page=4, slice=[0, 10]

Also start begins at 0, and page begins at 1.