TDX-8235098 Make Primo disjunction queries try to group words
Disjunctions in Primo aren't grouping keywords in meaningful ways.
This results in (A OR B) NOT C searching as A OR (B NOT C).

This change will group the queries inside the keywords list passed to
Primo.  Only if there are no fielded subtrees in the search.  This
is not a final solution, but a first draft that gathers the low-hanging
fruit.