Proxy the ACME challenge for Let's Encrypt
This is a temporary fix to a transitional problem.
Not every application server is behind the same reverse proxy
that answers the http challenge for Let's Encrypt.  So
rather than hope that the DNS catches before the failure limit
is reached, we are going to reverse proxy the challenge to the
one that does.

Which means we can't rely on DNS to resolve to the IP address.
And the ssl certificate doesn't include the IP address, so
we have can't verify the certificate.

It's wrapped in an environment variable to turn on and off
so that we don't run it on the host that gets proxied.